campaignd: Allow existing add-ons with invalid names to be reuploaded
This change is for backwards compatibility purposes only in order to
allow authors to continue uploading new versions of add-ons with
previously valid names that have been rendered invalid by the whitelist
addon_name_legal() method. This commit should be reverted next time an
add-ons server instance is deployed (hopefully for 1.12 first instead of
1.13.x).

A warning is provided in the logs for admins whenever an add-on with an
invalid name is reuploaded.