Add combo selection for campaign's entry points to mp::configure.
The selection is only available when there is more than one scenario
which belongs to campaign and has 'allow_new_Game' set to 'yes'.

The combo items are named using 'new_game_title' attribute from
[scenario] tag. If attribute does not exists, 'name' attribute is used
as a fallback.

This commit is for gui changes only. The entry point selection itself
is not yet functional.