New [campaign] tag attributes to specify the number of players.
The attributes are 'min_players' and 'max_players'. Their default value
is '2'. If 'max_players' is greater than 'min_players' - 'max_players'
would be equal to 'min_players'.

The number of players for campaigns is displayed similarly as for
stand-alone scenarios, except that if 'min_players' is not equal
'max_players', the number is displayed in 'min to max' format.