Added a check to change a unit's ellipse depending on its ZoC
This modification allows the game to automatically select the unit's ellipse, by checking if said unit emits a ZoC; if the unit loses or gains it, due for example to a level-up or to the Stun ability, the ellipse will change accordingly. This effectively makes no longer useful declaring the ellipse in a unit's .cfg file.