Renamed hero and leader nozoc ellipses to make them compatible with the automatic ellipse management
I had to rename these ellipses to make them compatible with the IS_HERO macro as well. This way, a user can assign a ellipse with the usual ellipse=misc/ellipse-hero, and when a unit loses its ZoC, it'll be updated accordingly. Leaving the file names as they were wasn't a good solution, because it required a way to split the ellipse-hero string to insert the nozoc- part.
With this solution, no WML change is required; however, we may want to add a wmllint rule to warn UMC authors if they have a ellipse= key in their unit cfgs.