ugly workaround for another MSVC compile error
The problem is that MSVC requires n in char buf[n] to be a constant; it
cannot be computed previously. It needs to be given as a number directly
(char buf[10];) or be a const var (const int n =10; char buf[n];)
It's ugly because of manual new/delete but other than that I'd have to
dive deeper into the code it seems. If the code is being worked on, please
remove the manual memory management and, for instance,
use std::string or stringstream for the get_map() callstack instead.