Reserve sides for players who played them in previous scenario.
The information for the previous scenario's sides which have "save_id"
and a player ("player_id") assigned to them are now being recorded in
"mp_game_settings". This information is used in mp::connect for every
non-last scenario to reserve sides for previous players so that
they could easily join the game and be placed in the correct side.