Have pump() no longer pump events raised while it is running.
The main way I know of for getting into this situation is hitting
the recursion maximum within pump(), and if that happens you probably
have runaway WML. So it should be better to move on instead of
continuing to process events.
(Normally, events raised within an event handler are pumped before
that event handler exits, so would not be affected by this change.)

I think the old behavior is an artifact of the old implementation,
but I'm making this a separate commit in case removing this causes
problems (easier to diagnose, especially if I'm not the one diagnosing).