Remove a .gitignore file that is not a .gitignore file
This thing was introduced in 492efa7b4bf687cada5d125d2fc4ab1c9ab097db
(from SVN r35965) due to a misapplication of a SVN property manipulation
command. I don't know what it is, but since it's part of the
decommissioned stats.w.o codebase and with the wrong filename to begin
with, I doubt anyone will miss it much.