addon/mg: Trap config::error exceptions thrown during .pbl read
This introduces a new minimal exception object to encapsulate some
relevant information (the file path and the config::error message) so it
can be propagated to the manager UI code correctly.

This is the first step is ensuring faulty .pbl files are not reported
during connection to the add-ons server as a "network communication
error".