Speed up AI turns 5-10% by deferring tod creation
Profiling --no-gui runs (I used a commercial tool) showed a surprising
amount of time is spent inside terrain_filter::match_internal. This
appears to mostly be as a result of checking every single terrain tile
to determine where the leader location is, which seems ripe for
optimization itself.

However, within this function 5-10% of the time is spent creating a
time_of_day that is rarely used by these AI matches. Simply reordering
the lines to defer its creation until inside the if block that checks
whether tod is a factor eliminates this part of the overhead.

This patch speeds AI turns up by 5-10% on my test machine (an 8-core
Linode 2048).