Add the function utils::might_contain_variables().
This function does a quick, but imprecise, check to see if a string
might change after variable interpolation. If this function returns
false, it is safe to assume the string will not change.

I thought I saw another function that does this, but I cannot find it.

The usecase I have in mind is where a string will persist for a while,
perhaps for an entire scenario, where optimizations could be made if
the string was constant, and where variables would be uncommon (with
"uncommon" relative to the size of the optimizations). For example,
event names seem to fit.