Use "Use map settings" to lock options in mp::connect.
A new "force_use_map_settings" attribute has been added to disallow
playing with custom settings in order to avoid broken gameplay
experience. The attribute's default value is "no", unless mp::connect is
is opened after winning a scenario. In that case, the attribute
determines whether "Use map settings" are in use or not.