Always use mp::connect_engine to set up an MP scenario.
The engine was slightly modified to have an option to force import
players when starting a game, so that previous players would be carried
over and correctly assigned to their corresponding sides.

As a side effect, the play_game() loop received a slight clean up,
because some code there became unnecessary.