Make it possible to overwrite a default leader by mp::connect selection.
Default leader is found by checking for "id" in [side] tag. If id was
found, it will check if leader was carried over and is now in [unit]
tag. If leader was still not found, it will try to pick a first unit
which can recruit.

If "Use map settings" are off or no leader was specified, mp::connect
will allow to select a leader from a faction. If default leader was in
carryover and the new leader was selected, the old leader attributes
will be cleared, keeping only an "id" and a newly set "type" and
"gender" attributes.