gui2/tgame_paths: New dialog for displaying game paths in Preferences
The button to invoke this dialog is currently in Preferences -> General.
This will change as soon as I think of a better place/someone else
suggests a better place for the button.

The goal is to allow users to quickly inspect the various paths used by
the game to locate resources and configuration, for debugging or content
creation purposes. The dialog provides options to copy directory paths
to the system clipboard and browse them using the platform's file
manager software, when available.