Add functionality to launch the platform's file manager for a path
This supports X11 platforms, Windows, and OS X. The code was
developed and tested on Debian GNU/Linux, Windows XP, and Windows 7.

The X11 and OS X versions of the function are identical and use an
exec/fork/wait pipeline with the only platform-dependant variable
being the external launcher binary (xdg-open on X11, open on OS X). I
opted for this somewhat ugly approach instead of system() because of
the many dangers resulting from allowing the system shell to
reinterpret the command line at whim.

The Win32 version uses ShellExecute() to launch the default
application associated with the Folder object type, in case Windows
Explorer is not the default handler for whatever reason.

The OS X code path was preliminary tested by mattsc and that's really
all I know about it. More testing on that platform would be very,
*very* welcome, although it is expected to behave pretty much the
same as Linux/X11