Updated "mp_game_settings" after MP scenario is played.
Now mp::lobby is able to display a correct scenario name, number of
turns and map_data for all scenarios in a campaign.

Some "mp_game_settings" are still not being updated to their correct
values after an MP scenario. For this to happen, mp::configure could be
splitted to gui and engine classes and the new engine class could be
used along with mp::connect_engine to update those values.