Improve i18n of system tray notifications on Windows
The system tray icon tooltip and some notification headers were not
translatable, which resulted in them being hardcoded to English. Fixed
this by making them translatable.

Caveat: the game_display::send_notification method is designed around
comparing some of the relevant strings. This means that some
functionality may not work properly if the translation language is
changed between notifications. It doesn't seem like anyone considered
this when designing the platform-neutral interface of this
functionality, though, seeing as how there's a call to
game_display::send_notification elsewhere that already uses a localized
argument.