addon/mg, gui2/taddon_uninstall_list: Implement reading of add-on titles from _info.cfg
This makes it so the user-facing parts of the UI display the actual
add-on titles from the add-ons server that were stored in the _info.cfg
files when installing. I had to move more code around than I wanted for
this because gui2::taddon_uninstall_list needs to handle not just add-on
ids and names, but also list item number mappings since GUI2 doesn't
allow attaching additional data to listbox items as far as I know.

It would be desirable to allow displaying the actual unedited paths or
ids for reference purposes for the user again (e.g. manual removal of
problematic add-ons) later, without disrupting the standard workflow.