variation_id replaces the former variation_name.
The new [variation] attribute variation_id is now used to identify the
subtype of the variation. This was done by the variation_name until now.
variation_name is now used to display a translated string to the user.

During creation of the unit tree a new attribute base_ids is stored to
hold a list of the [base_unit]s.