Turn wmi_container into a container.
This defines some types and functions that come from the STL
definition of a container. It's not complete, though, as I do
not see a reason to define size(), max_size(), or swap() at the
moment. (They would be easy to add, if the need ever arises.)

The new #include file might be useful elsewhere as well. I'll look
into that.