filesystem: Ensure SHGetSpecialFolderPathA call uses WINAPI call conv
gfgtdf reported on IRC that building Wesnoth on MSVC++ 2010 and
using a relative path to My Documents with the --config-dir command
line swich results in a runtime error. Paraphrased:

> The value of ESP was not properly saved across a function call

This is caused by mismatched function call conventions between
GetProcAddress()'s return value (__stdcall, a.k.a. WINAPI), and the
function pointer used to store and later call that address (__cdecl).

This commit should fix the issue by ensuring the function pointer is
set to use the WINAPI call convention.