Maintain difficulty in MP, as it is maintained in SP.
This solves the problem of always seeing a "Medium" difficulty in saves
preview pane for MP campaigns.

N.B. it should not be required to set difficulty in multiple places for
things to work correctly. In the future, it should be made so that
setting difficulty in "game_state::classification().difficulty" would be
enough.