Prevent crash of replay started from start-of-scenario save
Start-of-scenario saves in 1.11 do not contain the [replay_start] tag
any more because its contents are identical to the starting situation.
As a result, the tag needs to be created during the start-up sequence
of a scenario.  Whether this is necessary was previously decided by
checking whether the scenario snapshot was empty.  This works when the
scenario is started by finishing the previous scenario.  However, when
the start-of-scenario savegame is loaded, the snapshot is not empty.
Thus, we need to check whether the snapshot is empty OR whether the
turn number in it is less than 1.