Consider "Hold Position" units in endturn handling.
Do not show the dialogs for missed units if they are told to hold
position.
Changed the string in the dialog that asks about units which have not
been moved at all to be different to the one in the dialog that asks
about units which have still movement left.