Parameter consistency between can_show() and fire_event().
(These are members of wml_menu_item.)
I opted to get rid of a parameter and use a global. Using globals
is kind of yucky, but in this case, it's basically a choice between
using the global within the procedure or using the global as a
parameter to the procedure. I chose to go with the older function's
approach.