Restore support for negative healing.
The 1.10 implementation did not rule this out, so I've (finally)
tried to restore what 1.10 allowed, with slightly better accounting
of multiple sources of negative healing. I don't know if this is
the best thing to do, but some UMC authors wanted it back. (They
did not want it enough to answer my questions about it, though.)