Ignore disabled handlers when checking for existing ids and logging.
Before this update, you could have a one-time-only event with an id
fire, then attempt to add another event with that same id. Whether
or not the second event would stick depended on whether or not the
game was saved and reloaded between the first event firing and the
second event being added.