addon/mg: Check _info.cfg file existence before trying to open it
This is necessary to prevent spurious errors when preparing the Remove
Add-ons dialog such as:

> 20131110 02:21:21 error filesystem: Could not open '/home/shadowm/.wesnoth-1.11/data/add-ons/After_the_Storm/_info.cfg' for reading.

The installation status file is only used to read the original UI name
for the add-on from the add-ons server without establishing a network
connection. If the file doesn't exist we just carry on with an
automatic title generated via string manipulation, so the logged error
or lack thereof when the file does not exist is completely
inconsequential for our ends.