With shared pointers, we no longer need a deep copy.
Technically, this is a behavior change, since copies and assignments
will now share data, causing changes in one copy to affect the other.
However, the logic is that only one wmi_container is active at a time;
the copies and assignments are used to transfer data between scenarios
and to/from saves. So shallow copies are permissable (and slightly
reduce memory consumption).