When modifying a WML menu item, make a copy first.
This takes advantage of the implicit buffering provided by smart pointers.
If the menu item object is being used, that object will not be changed,
but instead will be destroyed when it is done. Eventually, this will
allow the explicit buffering of menu items to be eliminated, but there
is some functionality to shift around first.