Rename game_events::commit() to game_events::flush_messages().
With the removal of explicit WMI buffering (my previous commit),
flushing messages is the only thing this function does, so the
name change seemed to be in order. I've pulled this out as a
separate commit to draw attention to the call from menu_events.cpp.
That line was added in 50d4d5e31963f8b001c3dc32600d10d192519ee1
(f.k.a. r37758) with the log message "Ensured events registered on
command line are taken into account." Based on that message, it
seems like that line could now be removed, but I am not sure. So
I'll leave it in for now.