Set WML variable turn_number correctly in [pre]start events
There is a C++ function turn() which has value 1 during prestart and
start events.  The same is the case for wesnoth.current.turn in Lua.
By contrast, WML variable turn_number was not updated until after the
start event, resulting in it either being undeclared or containing
whatever the previous value was (such as the last turn number of the
previous scenario).  Now it behaves the same as turn() and
wesnoth.current.turn.