SotBE S16: eliminate dependence on luck at scenario start
The difficulty of the scenario depended too much on the success of a
small number of attacks, on whether the player succeeded in killing
one, or even both, of the close enemy leaders on the first turn.
Now, Kapou'e has his own keep, out of reach of the two close enemy
keeps.  Minor changes going along with this are making the two AI
leaders warriors (instead of grunts) and not allowing wolves as
replacements for Jetto and Inarix.