New lua metatable "wesnoth.mp_settings". Follows pattern of previous lua hooks in src/scripting/lua.cpp.
I would like to be able to access all the fields that occur in
"multiplayer" tag in save game files.

Referring to gamestatus.cpp line 1208, which is what writes
the game state to a config for output, I check if

(game_state.classification().campaign_type=="multiplayer" ),

and if so I ask the mp_settings object for fields.

Referring to line 195 of mp_game_settings.cpp, I use the same names
as appear in save files for the lua hooks.

I don't currently handle vector objects like "active_mods".