Added new CampaignWML attribute "allow_era_choice".
If attribute is set to "no" and campaign with it is selected, eras menu will be
hidden and text to indicate this will be displayed.

Attribute defaults to "yes" and has been added to LoW.

N.B. since current MP code is tighly coupled with eras, the default era will
still be used by the engine to properly initialize stuff. Later on, code should be
improved by making it more modular and independant of eras.