NR 05a - The Pursuit: add dialogue when a unit steps on the dungeon signpost
Most of the other signposts elicit a remark when a unit steps on them, but
curiously not this one.

Not only do I give the discovering unit a line, I have Tallin speak as
well. This may give alert first-time NR players a clue that this corridor
is important.

Some may prefer that players continue stumbling about, and object.
Unfortunately, there's no NR maintainer or active lord of prose to resolve
any difference of opinion.