campaignd: Abort sending if the pack size is negative (bug #20893)
Sending an add-on whose size could not be determined means that the pack
in which it is stored is inaccessible in some way. Proceeding in this
situation, with the network::send_file() fix for the same condition,
results in the built-in add-ons client stalling forever waiting for
campaignd to send data to it, so we must respond with an [error] tag
instead.

This is the final step in fixing bug #20893.