HttT: remove obsolete "wmllint: recognize" magic comments
The exception is Gryphon Mountain, where the magic comments are updated to
recognize the awakened ids.

This is not the only approach that could be used. wmllint could be upgraded
to recognize the transformed ids in [modify_unit]. This is more involved
than I want to do right now, for what I expect to be a corner case.

The other way would be to also filter the awakening events by type as well
as id, and not change the id at all. The original awakening code killed the
old unit and created a new one in its place, and the new id distinguished
the new unit from the old one. However, now that we are modifying an existing
unit, we do not need to change ids, especially since this is not seen by
the player.

For now, though, the easiest path is just to tweak the already existing
magic comments.