about: Split multi-line section headers for the credits screen
Multi-line section headers such as those generated by campaigns with
multi-line titles result in slight scrolling glitches because the
scrolling code assumes every entry in the text vector to comprise a
single line.

We don't want to split those headers for the help screen because the
current approach results in nicely formatted headers for the
Contributors page there, so a hacky compromise for now should be to only
split multiline headers for the scrolling credits.

An alternative would have been to make the text vector a linked list so
about::show_about could deal with this issue directly, but that seems
rather wasteful.