display: Drop trailing underscore from the color_adjust_blue_ WML option
This was used in saved games during the serialization of the display
configuration (see the display::write() method) since around 1.10.x. The
trailing underscore was clearly not intentional, as the siblings of this
attribute for the red and green channels don't have it, so drop it.

Resulting from this commit, saved games from previous versions that did
have the trailing underscore and were affected by a [color_adjust]
action may be loaded with the wrong color channel configuration (blue
will be reset to the default of zero), but this situation should be rare
enough to not absolutely require backwards-compatibility code.