NR 05a: change "Should this unit" to "Should I" in Rod dialog
The other choice dialogs in this scenario use "I" or "we", so
this brings it into conformity with them.

On the mailing list, I also spoke of a desire to have Camerin's
lines run before the choice is made, but looking at it, this
is a more complex change than I want to deal with right now.