campaign: Implement [server_info] block and feedback_url_format attribute
The [server_info] configuration block is intended to contain additional
information about a campaignd instance that may be used by clients or
the server for various purposes.

As a first such information field, the feedback_url_format string
contains a format string for use with
fast_interpolate_variables_into_string() to build add-on feedback page
URLs.