campaignd: Implement .pbl [feedback] block and [campaign] feedback_url field
The .pbl [feedback] block (also used in the server configuration WML)
specifies a list of format values interpolated into the
feedback_url_format string from [server_info] to build a web page URL
that is delivered to clients each time an add-on's metadata is sent in a
[campaign] block.

For example, given the following campaignd-side configuration:

  [server_info]
      feedback_url_format="http://forums.wesnoth.org/viewtopic.php?t=$thread_id"
  [/server_info]

And the following tag in the _server.pbl WML:

  [feedback]
      thread_id=32091
  [/feedback]

..., the server will deliver the following field as part of the
[campaign] block provided in the response to [request_campaign_list]:

  feedback_url="http://forums.wesnoth.org/viewtopic.php?t=32091

The fields from the [feedback] block are percent-encoded to ensure that
the .pbl WML can't manipulate query strings from the
feedback_url_format. If for a system administrator it is somehow
desirable to let authors enter arbitrary URL fragments (e.g.
feedback_url_format="$url" and [feedback]
url="http://this-site-kills-kittens.example.com, bad idea btw), then
they should probably remove the utils::urlencode() call themselves and
take responsibility for any malicious links that could be served to
users of the add-ons server that way.