gui2/taddon_description: Completely redesigned dialog layout
The new layout merges some information rows together to increase
available vertical space for displaying additional fields later, and
uses a 'description' scroll_label type for the add-on Description field
to improve its appearance, especially for multi-line descriptions.

This also makes the dialog non-click-dismissable and adds a Close button
on the bottom right corner. Since we're about to add more interactive
controls, there's really not much of a choice about the click-dismiss
option; it gets disabled for such dialogs automatically.