SotBE 04: fix ORCISH_SHAMAN id recognition
While I tested my whofield feature, unfortunately it was on an
old version of the campaign in which all shamans were on side 1,
so the macros had no side field. Recently, it was decided that
the shamans in this scenario should start on side 3, and a side
field was inserted into the macros.

When it was time to commit, there was of course a conflict, but
I simply put my existing magic comments into the new version,
not considering that now the id field had been shifted to the
right by one, and didn't re-run wmllint.