SotBE 17: recognize speaker Bruce
While I looked for "unknown speakers" in an old version of the campaign,
in the meantime mattsc had changed the scenario to use a macro for the
enemy reinforcements (e5905f05).

Although I could use my spiffy new "whofield" magic comment here, Bruce
is the only general who speaks, so an old-fashioned "recognize" is more
appropriate.