New wesnoth-defaults utility script
I wrote this a long time ago and have been using it to ease testing of
Wesnoth's default configuration under different conditions ever since.

It creates a new configuration and data directory under /tmp that only
exists for the duration of the Wesnoth session, and disables WML cache
functionality.

This is extremely useful for quickly testing and debugging functionality
dependant on user configuration settings.