campaignd: Don't prefix errors sent to clients with "#Error: "
The string only worked as intended before the GUI2/Pango days (i.e.
before some 1.5.x version), meaning that the client would use a red
color for the font.

In other words, deplorable user interface design.

Nowadays, the add-ons client uses a dedicated error message path for
presenting this kind of alerts to the user, which uses a dialog with a
visible "Error" caption and everything. We don't really need extra
formatting, and in any case, *it should not be the server's mission to
do this*!