Fix logic of [filter_vision] in Standard Unit Filters (SUFs)
Previously, the [filter_vision] tag would cause a false positive when a
unit was on terrain on which it was invisible and that also was under
fog/shroud.  There might have been other combinations that did not work
either, I did not check that.  I did check that all 8 combinations of
visble=yes/no, the unit hiding due to terrain or not, and the hex being
under fog/shroud or not work now.