Add missing check for a null surface from image::get_image()
Fixes a NULL pointer dereference when drawing preview pane overlays in
the Recall Unit dialog for a unit whose overlays list includes an image
that doesn't exist or can't be read, and debug mode is not enabled.

This didn't happen in debug mode because image::get_image() retrieves a
valid placeholder surface (currently images/misc/missing-image.png) in
that situation with debug mode enabled.

Reported by SkyOne in the forums: http://r.wesnoth.org/t39983

Thanks to mattsc for pinpointing the commit at fault.

Note: the code handling overlays on the listbox/menu of the Recall Unit
dialog shouldn't be affected by this since it uses the BLIT image path
function instead of accessing surfaces directly.