campaign: Remove convert_binary_to_gzip() method and related functionality
This has been around since commit 727294a4b5c72551f559645de504d3a9b8cdf49d
(version 1.5.1) and doesn't serve any purpose nowadays since add-ons
are always stored in gzip format and with the correct encoding.

The method was apparently repurposed a few times to deal with some
encoding issues that cropped up during the introduction of server-side
gzip streams in 1.5.x:

 * fdb32be49eb17023364ed2df3d4e60bf582082db "Fixed campaign to use
   correct compression level and encoding messages"
 * ed564762873d3d8ad043acb720112f92cf848ca3 "Fixed addon server
   conversion code for cr encoding"
 * f4c62eabe7e4d97eaf4a46edc0f67d75631e2cd1 "Made campaignd server
   encode CR in old content that is stored in campaignd"
 * efab5f83008b1f9937b1c101a3c1eaff87fd3fd8

All in all, the last time it possibly saw any use for wesnoth.org's
usage was with version 1.6.0, therefore it's safe to get rid of it now.

(The encoded and cr_encoded attributes in the server configuration files
for 1.8, 1.9/1.10, and 1.11 in production have been set to 'yes' for
ages already, which means the method wasn't being run beyond the first
statement on startup anyway.)