Fix drawing the offmap area for small resolutions.
The problem was a bit tricky. A partial resolution removed an element with
an empty id. The main_map_border section has no id and thus an empty id.
So the inheritance removed this section by accident. Of course the id for
a removal or change should be mandatory, which the next patch will do.

This fixes a part of bug #21584.