Disallow change and remove sections without an id.
This change to the ThemeWML will avoid the offmap drawing issues due to
accidentaly changing or removing the main_map_border section from a
resolution. If it is really wanted it is possible to add an id to this
section, the game doesn't use it, but inheritance uses the raw config.

This fix rewrites the expand_partialresolution, adding the check and
making the code easier to follow.

This should avoid bugs like bug #21584.