Add a work-around for a blurring issue.
When pressing the next or previous tip button there is a drawing glitch.
As soon as the mouse pointer leaves the button area the glitch
disappears. For now force an update directly after the button is
pressed. This is not particularly elegant, but it works. The SDL2
transition will probably fix this issue automatically.