Optimizing game event firing.
This uses the indexing of handlers by name to reduce the number of
event handlers that are checked when an event is fired. There is
extra overhead to independently track handlers whose names contain
variables, but in more common cases, there should be an improvement,
particularly when many events fire in sequence, such as the enter/
exit_hex events.