HttT S17: Add a wmllint: noconvert directive amidst Lua code
wmllint doesn't know what Lua is or that it should skip it.

The code in question was triggering the viewing_side -> side plain text
substitution rule because wesnoth.find_path takes an optional last table
argument that includes a 'viewing_side' table entry, that is completely
unrelated to the [filter_vision] viewing_side -> SSF change in 1.11.x.