gui2/ttext_box: Workaround text AA issue with transparent background
... by making it not transparent.

For some reason, on Linux the antialiasing of the textbox text contents
looks slightly off when the background is a transparent (a != 255)
color. This does not happen when a solid color (a == 255) is used
instead.

Since Cairo AA is broken horribly with ClearType on Windows atm I'm not
really sure this specific bug also happens there.

To keep the intended effect of ebc214ff124ad4bb79700b509b1369766481c842,
I took a couple of screenshots of the textbox widget in both its enabled
and disabled states and averaged the background colors with the GIMP to
get an approximation of the result of blending the transparent overlay
from the GUI__BACKGROUND_COLOR_* macros with the standard dialog
background.