Rename some variables to reduce confusion.
The variable 'n' was declared within a scope that had already
declared 'n'. While each variable is only used in the line in which
it is defined (they are loop control variables for 'for' statements),
it's just bad style to have that potential confusion.

Plus, 'n' is not a descriptive name. (Not that the new names are
much better.)