Turn dialog used for hotseat mp games now applies a blindfold
This is intended to address bug #3856.

Following discussion on IRC, this is considered a feature request
and not a bug, and so cannot be addressed during feature freeze
for 1.12. Hence I am pushing it to a topic branch.

I can't push a changelog change either as we don't know what version
this will go in, but the changelog entry should be something like:

"Fix bug #3856: The turn dialog used in hotseat MP play now applies
a blindfold to the map for the duration of the dialog."