Make all labels visible in the editor.
I don't know why the earlier logic was so convoluted. Someone tried
to mash so many conditions in one line that apparently I missed the
intent earlier. I'm no longer trying to match the earlier logic, but
instead just trying to get the conditions right.

I really don't know how this will interact with the map editor's
label tool, because that tool is disabled for me and the screen
gives me no indication of how to enable it. But this does appear
to fix bug #21732 without messing up multiplayer.