Move window caption initialisation code.
The code is moved from game_controller::init_language to
game_controller::init_video. This makes porting the code to SDL 2
easier. In SDL 2 the caption must be set after the main window exists.
It could be set in the constructor of the Window, but that can be done
once we drop SDL 1.2 support.